bugfigx - fixing segfaults for large sequences

+++ pysam/csamtools.pxd
@@ -239,8 +239,8 @@ cdef extern from "pysam_util.h":
 
     # add *nbytes* into the variable length data of *src* at *pos*
     bam1_t * pysam_bam_update( bam1_t * b, 
-                               uint8_t nbytes_old,
-                               uint8_t nbytes_new,
+                               size_t nbytes_old,
+                               size_t nbytes_new,
                                uint8_t * pos )
 
     # translate char to unsigned char
